requirements.txt
# Core requirements
tqdm>=4.65.0
numpy>=1.24.0
pandas>=1.5.3
transformers>=4.30.0
torch>=1.10.0
ijson>=3.2.0
accelerate>=0.20.0
einops>=0.6.1
bitsandbytes>=0.41.0



# New requirements for enhanced experiments
scikit-learn>=1.0.2
matplotlib>=3.5.0
seaborn>=0.11.2
pyyaml>=6.0
rich>=13.0.0
tensorboard>=2.8.0
python-dotenv>=0.19.0
psutil>=5.9.0
# faiss-gpu removed

# RAG-Fusion specific requirements
rank_bm25>=0.2.2
sentence-transformers>=2.2.2

# Additional packages
openai>=0.27.0
loguru>=0.6.0

